+++ packages/styled-system/src/index.ts
@@ -1,7 +1,6 @@
 export * from "./config"
 export * from "./css"
 export * from "./system.types"
-export * from "./theming.types"
 export * from "./system"
 export * from "./create-theme-vars"
 export * from "./pseudos"
@@ -13,3 +12,4 @@ export type {
 } from "./utils"
 export { tokenToCSSVar } from "./utils/create-transform"
 export type OmitSpaceXY<T> = Omit<T, "spaceX" | "spaceY">
+export type { ThemeTypings, BaseThemeTypings, CustomThemeTypings } from "./theme.types"