+++ packages/styled-system/src/theming.types.ts
@@ -1,30 +1,6 @@
 /**
  * This file will be replaced by running "npx @chakra-ui/cli tokens"
  */
+import type { BaseThemeTypings } from "./theme.types"
 
-export interface ThemeTypings extends EmptyThemeTypings {}
-
-export interface EmptyThemeTypings {
-  borders: string
-  colors: string
-  breakpoints: string
-  colorSchemes: string
-  fonts: string
-  fontSizes: string
-  fontWeights: string
-  layerStyles: string
-  letterSpacings: string
-  lineHeights: string
-  radii: string
-  shadows: string
-  sizes: string
-  space: string
-  textStyles: string
-  zIndices: string
-  components: {
-    [componentName: string]: {
-      sizes: string
-      variants: string
-    }
-  }
-}
+export interface ThemeTypings extends BaseThemeTypings {}