+++ examples/nextjs-typescript/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Change Log
 
+## 1.1.13
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chakra-ui/react@1.8.6
+
+## 1.1.12
+
+### Patch Changes
+
+- [`e1fe48cbe`](https://github.com/chakra-ui/chakra-ui/commit/e1fe48cbe37324744cfe6184d785c093cda1125e)
+  Thanks [@TimKolberger](https://github.com/TimKolberger)! - Bumped patch
+  version for every package to fix release process. Root cause was a bug in our
+  CI configuration.
+- Updated dependencies
+  [[`e1fe48cbe`](https://github.com/chakra-ui/chakra-ui/commit/e1fe48cbe37324744cfe6184d785c093cda1125e)]:
+  - @chakra-ui/icons@1.1.7
+  - @chakra-ui/react@1.8.5
+  - @chakra-ui/theme-tools@1.3.6
+
+## 1.1.11
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @chakra-ui/react@1.8.4
+  - @chakra-ui/theme-tools@1.3.5
+  - @chakra-ui/icons@1.1.6
+
 ## 1.1.10
 
 ### Patch Changes