+++ packages/editable/CHANGELOG.md
@@ -1,5 +1,51 @@
 # Change Log
 
+## 1.4.0
+
+### Minor Changes
+
+- [#4443](https://github.com/chakra-ui/chakra-ui/pull/4443)
+  [`fbe946223`](https://github.com/chakra-ui/chakra-ui/commit/fbe94622357e22acaf8bab0eae33ceae663d7a5b)
+  Thanks [@heozeop](https://github.com/heozeop)! - Added the component
+  `EditableTextarea` to `Editable`. Use the textarea element to handle multi
+  line text input in an editable context.
+
+  ```tsx live=false
+  <Editable defaultValue="Change me" onChange={console.log}>
+    <EditablePreview />
+    <EditableTextarea />
+  </Editable>
+  ```
+
+### Patch Changes
+
+- Updated dependencies
+  [[`5cd5cff35`](https://github.com/chakra-ui/chakra-ui/commit/5cd5cff35e4837539d83a2157a07585d461b0aac)]:
+  - @chakra-ui/hooks@1.8.5
+
+## 1.3.7
+
+### Patch Changes
+
+- [`e1fe48cbe`](https://github.com/chakra-ui/chakra-ui/commit/e1fe48cbe37324744cfe6184d785c093cda1125e)
+  Thanks [@TimKolberger](https://github.com/TimKolberger)! - Bumped patch
+  version for every package to fix release process. Root cause was a bug in our
+  CI configuration.
+- Updated dependencies
+  [[`e1fe48cbe`](https://github.com/chakra-ui/chakra-ui/commit/e1fe48cbe37324744cfe6184d785c093cda1125e)]:
+  - @chakra-ui/hooks@1.8.4
+  - @chakra-ui/react-utils@1.2.3
+  - @chakra-ui/utils@1.10.4
+
+## 1.3.6
+
+### Patch Changes
+
+- Updated dependencies
+  [[`a870e6b94`](https://github.com/chakra-ui/chakra-ui/commit/a870e6b94367b7c6448d5c5c5aa8577e33e15e3a)]:
+  - @chakra-ui/utils@1.10.3
+  - @chakra-ui/hooks@1.8.3
+
 ## 1.3.5
 
 ### Patch Changes