+++ packages/popover/CHANGELOG.md
@@ -1,5 +1,44 @@
 # Change Log
 
+## 1.11.7
+
+### Patch Changes
+
+- [#5623](https://github.com/chakra-ui/chakra-ui/pull/5623)
+  [`5cd5cff35`](https://github.com/chakra-ui/chakra-ui/commit/5cd5cff35e4837539d83a2157a07585d461b0aac)
+  Thanks [@Brennvo](https://github.com/Brennvo)! - Fixed an issue where the prop
+  `isLazy` did not work as expected. This was achieved by updating the hook
+  `useAnimationState`.
+- Updated dependencies
+  [[`5cd5cff35`](https://github.com/chakra-ui/chakra-ui/commit/5cd5cff35e4837539d83a2157a07585d461b0aac)]:
+  - @chakra-ui/hooks@1.8.5
+
+## 1.11.6
+
+### Patch Changes
+
+- [`e1fe48cbe`](https://github.com/chakra-ui/chakra-ui/commit/e1fe48cbe37324744cfe6184d785c093cda1125e)
+  Thanks [@TimKolberger](https://github.com/TimKolberger)! - Bumped patch
+  version for every package to fix release process. Root cause was a bug in our
+  CI configuration.
+- Updated dependencies
+  [[`e1fe48cbe`](https://github.com/chakra-ui/chakra-ui/commit/e1fe48cbe37324744cfe6184d785c093cda1125e)]:
+  - @chakra-ui/close-button@1.2.7
+  - @chakra-ui/hooks@1.8.4
+  - @chakra-ui/popper@2.4.3
+  - @chakra-ui/react-utils@1.2.3
+  - @chakra-ui/utils@1.10.4
+
+## 1.11.5
+
+### Patch Changes
+
+- Updated dependencies
+  [[`a870e6b94`](https://github.com/chakra-ui/chakra-ui/commit/a870e6b94367b7c6448d5c5c5aa8577e33e15e3a)]:
+  - @chakra-ui/utils@1.10.3
+  - @chakra-ui/close-button@1.2.6
+  - @chakra-ui/hooks@1.8.3
+
 ## 1.11.4
 
 ### Patch Changes